+++ src/Sql/Sql/DataClassification/Cmdlet/EnableAzSqlInstanceDatabaseSensitivityRecommendation.cs
@@ -57,7 +57,7 @@ namespace Microsoft.Azure.Commands.Sql.DataClassification.Cmdlet
 
         protected override ManagedDatabaseSensitivityClassificationModel PersistChanges(ManagedDatabaseSensitivityClassificationModel entity)
         {
-            ModelAdapter.EnableManagedDatabaseSensitivityRecommendations(entity);
+            ModelAdapter.EnableSensitivityRecommendations(entity);
             return null;
         }
     }